Add vsock cli param
Using an extra CLI parameter, we are able to specify a unix socket
to be used as a VSOCK endpoint for Guest-to-Host communication.

For now, the argument is just the socket path, with a hardcoded
Guest ID. The default value is empty so that the VSOCK endpoint
is not added by default.

Example invocation:

./dbs-cli --kernel-path ./vmlinux \
	  --rootfs ./rootfs.img \
	  --boot-args 'console=ttyS0 tty0 pci=off root=/dev/vda' \
	  --vsock vsock.sock

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>