cli: add api server
api server and api client are added to the dbs-cli. Through API server,
users could send commands when VM is running, so that we could use
dbs-cli to test operations like device hotplug.

Besides, since api_server and cli_instance share lots of common
functions so a new trait VmmComm is introduced to decreate the
repetitive rate of the same code.
More information could be found in README file and args.md.

Signed-off-by: Chao Wu <chaowu@linux.alibaba.com>